LPS-118935 Move to display package, removing from either class is a breaking change.
