+++ modules/apps/change-tracking/change-tracking-spi/bnd.bnd
@@ -3,7 +3,6 @@ Bundle-SymbolicName: com.liferay.change.tracking.spi
 Bundle-Version: 1.1.0
 Export-Package:\
 	com.liferay.change.tracking.spi.display,\
-	com.liferay.change.tracking.spi.display.base,\
 	com.liferay.change.tracking.spi.display.context,\
 	com.liferay.change.tracking.spi.exception,\
 	com.liferay.change.tracking.spi.listener,\
