--- modules/apps/change-tracking/change-tracking-spi/src/main/resources/com/liferay/change/tracking/spi/display/base/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
